refactor: split ACP built-in plugin into focused modules
- keep acp-plugin.ts as composition/registration entry

- extract config, context, parsers, client, store, service, and tools

- preserve behavior while improving maintainability
